+++ src/index.js
@@ -38,7 +38,7 @@ module.exports = {
         'jsx-a11y/mouse-events-have-key-events': 2,
         'jsx-a11y/no-access-key': 2,
         'jsx-a11y/no-onchange': 2,
-        'jsx-a11y/onclick-hs-focus': 2,
+        'jsx-a11y/onclick-has-focus': 2,
         'jsx-a11y/onclick-has-role': 2,
         'jsx-a11y/role-has-required-aria-props': 2,
         'jsx-a11y/role-supports-aria-props': 2,