+++ tests/src/rules/aria-role.js
@@ -55,6 +55,8 @@ ruleTester.run('aria-role', rule, {
     { code: '<div role={role || "button"} />', parserOptions },
     { code: '<div role={role || "foobar"} />', parserOptions },
     { code: '<div role="tabpanel row" />', parserOptions },
+    { code: '<div role="doc-abstract" />', parserOptions },
+    { code: '<div role="doc-appendix doc-bibliography" />', parserOptions },
     { code: '<Bar baz />', parserOptions }
   ].concat(validTests),
 
@@ -65,6 +67,7 @@ ruleTester.run('aria-role', rule, {
     { code: '<div role=""></div>', errors: [ errorMessage ], parserOptions },
     { code: '<div role="tabpanel row foobar"></div>', errors: [ errorMessage ], parserOptions },
     { code: '<div role="tabpanel row range"></div>', errors: [ errorMessage ], parserOptions },
+    { code: '<div role="doc-endnotes range"></div>', errors: [ errorMessage ], parserOptions },
     { code: '<div role />', errors: [ errorMessage ], parserOptions },
     { code: '<div role={null}></div>', errors: [ errorMessage ], parserOptions }
   ].concat(invalidTests)